+++ back-end/src/app.ts
@@ -9,7 +9,7 @@ import * as cookieParser from "cookie-parser";
 import { passportLocalStrategy } from "./util/user/passportLocalStrategy";
 import { Express, Request, Response, NextFunction } from "express";
 import { Connection, createConnections, ConnectionOptions } from "typeorm";
-import { errorHandler, error404Handler } from "./util/errorHandling/errorHandler";
+import { errorHandler, error404Handler } from "./util/errorHandling";
 import { mainRouter } from "./routes";
 
 // Load environment variables from .env file
@@ -52,10 +52,6 @@ export function buildApp(callback: (app: Express, err?: Error) => void): void {
 const expressSetup = (): Express => {
   // Create Express server
   const app = express();
-
-  // view engine setup
-  app.set("views", path.join(__dirname, "views"));
-  app.set("view engine", "ejs");
 
   // Express configuration
   app.set("port", process.env.PORT || 3000);
@@ -132,10 +128,7 @@ const createDatabaseOptions = (): ConnectionOptions[] => {
     password: process.env.DB_PASSWORD,
     database: process.env.DB_DATABASE,
     entities: [
-      __dirname + "/db/entity/user{.js,.ts}",
-      __dirname + "/db/entity/event{.js,.ts}",
-      __dirname + "/db/entity/hardwareItem{.js,.ts}",
-      __dirname + "/db/entity/reservedHardwareItem{.js,.ts}"
+      __dirname + "/db/entity/hub/*{.js,.ts}"
     ],
     // Per TypeOrm documentation, this is unsafe for production
     // We should instead use migrations to change the database
@@ -151,7 +144,7 @@ const createDatabaseOptions = (): ConnectionOptions[] => {
     password: process.env.APP_DB_PASSWORD,
     database: process.env.APP_DB_DATABASE,
     entities: [
-      __dirname + "/db/entity/applicationUser{.js,.ts}"
+      __dirname + "/db/entity/applications/*{.js,.ts}"
     ],
     synchronize: false,
     logging: false,