+++ back-end/src/controllers/index.ts
@@ -1,4 +1,5 @@
+export * from "./homeController";
+export * from "./achievementsController";
 export * from "./userController";
 export * from "./scheduleController";
-export * from "./homeController";
 export * from "./hardwareController";