+++ back-end/src/controllers/userController.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from "express";
 import * as passport from "passport";
-import { ApiError } from "../util/errorHandling/apiError";
-import { HttpResponseCode } from "../util/errorHandling/httpResponseCode";
+import { ApiError } from "../util/errorHandling";
+import { HttpResponseCode } from "../util/errorHandling";
 import { AuthLevels } from "../util/user";
 
 /**
@@ -28,7 +28,6 @@ export class UserController {
         } else {
           res.redirect("/");
         }
-        // res.send({ message: "Logged in" });
       });
     })(req, res);
   }