+++ back-end/src/db/entity/index.ts
@@ -1,5 +1,0 @@
-export * from "./user";
-export * from "./applicationUser";
-export * from "./event";
-export * from "./hardwareItem";
-export * from "./reservedHardwareItem";