+++ back-end/src/util/cache/abstract-classes/cacheObject.ts
@@ -3,9 +3,9 @@
  */
 export abstract class CacheObject {
   /**
-   * The id of the object in the database
+   * The id of the object
    */
-  public id: number;
+  public id: string;
 
   /**
    * The amount of time the object stays synced (miliseconds).
@@ -21,9 +21,9 @@ export abstract class CacheObject {
 
   /**
    * Creates a new basic cache object
-   * @param id The id of the object in the database
+   * @param id The id of the object
    */
-  constructor(id: number) {
+  constructor(id: string) {
     this.id = id;
     this.syncedAt = Date.now();
   }