+++ back-end/src/util/user/authorization.ts
@@ -1,8 +1,6 @@
 import { Request, Response, NextFunction } from "express";
 import { AuthLevels } from "./authLevels";
-import { User } from "../../db/entity";
-import { ApiError } from "../errorHandling/apiError";
-import { HttpResponseCode } from "../errorHandling/httpResponseCode";
+import { User } from "../../db/entity/hub";
 
 /**
  * Checks if the request's sender is logged in