+++ back-end/src/util/user/userValidation.ts
@@ -1,6 +1,8 @@
-import { getRepository, getConnection, Equal, AdvancedConsoleLogger } from "typeorm";
+import { getConnection } from "typeorm";
 import * as pbkdf2 from "pbkdf2";
-import { ApplicationUser, User } from "../../db/entity/";
+import { User } from "../../db/entity/hub";
+import { ApplicationUser } from "../../db/entity/applications";
+import { ApiError, HttpResponseCode } from "../errorHandling";
 
 /**
  * We check that the password hash is valid
@@ -66,7 +68,7 @@ async function getPasswordFromHub(submittedEmail: string): Promise<string> {
       return user.password;
     return undefined;
   } catch (err) {
-    throw new Error(`Lost connection to database (hub)! ${err}`);
+    throw new ApiError(HttpResponseCode.INTERNAL_ERROR, `Lost connection to database (hub)! ${err}`);
   }
 }
 
@@ -85,7 +87,7 @@ export async function getUserByIDFromHub(submittedID: number): Promise<User> {
 
     return user;
   } catch (err) {
-    throw new Error(`Lost connection to database (hub)! ${err}`);
+    throw new ApiError(HttpResponseCode.INTERNAL_ERROR, `Lost connection to database (hub)! ${err}`);
   }
 }
 
@@ -104,7 +106,7 @@ export async function getUserByEmailFromHub(submittedEmail: string): Promise<Use
 
     return user;
   } catch (err) {
-    throw new Error(`Lost connection to database (hub)! ${err}`);
+    throw new ApiError(HttpResponseCode.INTERNAL_ERROR, `Lost connection to database (hub)! ${err}`);
   }
 }
 
@@ -122,7 +124,7 @@ export async function getUserByEmailFromApplications(submittedEmail: string): Pr
       .getOne();
     return applicationUser;
   } catch (err) {
-    throw new Error(`Lost connection to database (applications)! ${err}`);
+    throw new ApiError(HttpResponseCode.INTERNAL_ERROR, `Lost connection to database (applications)! ${err}`);
   }
 }
 
@@ -141,6 +143,6 @@ export async function insertNewHubUserToDatabase(hubUser: User): Promise<void> {
       .execute();
     return;
   } catch (err) {
-    throw new Error(`Lost connection to database (applications)! ${err}`);
+    throw new ApiError(HttpResponseCode.INTERNAL_ERROR, `Lost connection to database (applications)! ${err}`);
   }
 }